requirements.txt
flake8
Django
djangorestframework
Pillow
python-decouple
django-filter
geopy